+++ noodles-vcf/examples/vcf_view_async.rs
@@ -20,7 +20,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .map(BufReader::new)
         .map(vcf::AsyncReader::new)?;
 
-    let header = reader.read_header().await?.parse()?;
+    let header = reader.read_header().await?;
 
     let mut writer = vcf::AsyncWriter::new(io::stdout());
     writer.write_header(&header).await?;